adding new component for carousel - search button now works
