+++ package.json
@@ -39,6 +39,7 @@
     "react": "^17.0.2",
     "react-bootstrap": "^1.5.2",
     "react-dom": "^17.0.2",
+    "react-tooltip": "^4.2.19",
     "react-use-hoverintent": "^1.2.3",
     "reactstrap": "^8.9.0",
     "sequelize": "^6.6.2"
